Remove platform-specific elements from core/dir.c and core/file.c
