+++ src/platform/file_manager.c
@@ -1,0 +1,247 @@
+#include "file_manager.h"
+
+#include "core/file.h"
+#include "core/log.h"
+#include "core/string.h"
+#include "platform/android/android.h"
+#include "platform/vita/vita.h"
+
+#include <dirent.h>
+#include <stdlib.h>
+#include <string.h>
+#include <sys/stat.h>
+
+#ifdef _WIN32
+#include <windows.h>
+
+#define fs_dir_type _WDIR
+#define fs_dir_entry struct _wdirent
+#define fs_dir_open _wopendir
+#define fs_dir_close _wclosedir
+#define fs_dir_read _wreaddir
+#define dir_entry_name(d) wchar_to_utf8(d->d_name)
+#define dir_entry_close_name(n) free((void*)n)
+
+static const char *wchar_to_utf8(const wchar_t *str)
+{
+    int size_needed = WideCharToMultiByte(CP_UTF8, 0, str, -1, NULL, 0, NULL, NULL);
+    char *result = (char *) malloc(sizeof(char) * size_needed);
+    WideCharToMultiByte(CP_UTF8, 0, str, -1, result, size_needed, NULL, NULL);
+    return result;
+}
+
+static wchar_t *utf8_to_wchar(const char *str)
+{
+    int size_needed = MultiByteToWideChar(CP_UTF8, 0, str, -1, NULL, 0);
+    wchar_t *result = (wchar_t *) malloc(sizeof(wchar_t) * size_needed);
+    MultiByteToWideChar(CP_UTF8, 0, str, -1, result, size_needed);
+    return result;
+}
+
+#else // not _WIN32
+#define fs_dir_type DIR
+#define fs_dir_entry struct dirent
+#define fs_dir_open opendir
+#define fs_dir_close closedir
+#define fs_dir_read readdir
+#define dir_entry_name(d) ((d)->d_name)
+#define dir_entry_close_name(n)
+#endif
+
+#ifndef S_ISLNK
+#define S_ISLNK(m) 0
+#endif
+
+#ifndef S_ISSOCK
+#define S_ISSOCK(m) 0
+#endif
+
+#ifdef __vita__
+#define CURRENT_DIR VITA_PATH_PREFIX
+#elif defined(_WIN32)
+#define CURRENT_DIR L"."
+#else
+#define CURRENT_DIR "."
+#endif
+
+#ifdef _MSC_VER
+#include <direct.h>
+#define chdir _chdir
+#elif !defined(__vita__) || !defined(__ANDROID__)
+#include <unistd.h>
+#endif
+
+int platform_file_manager_get_directory_contents_by_extension(char **file_list, int *file_count, const char *extension, int max_files)
+{
+#ifdef __ANDROID__
+    if (!android_get_directory_contents_by_extension(file_list, file_count, extension, max_files)) {
+        return 0;
+    }
+#else
+    fs_dir_type *d = fs_dir_open(CURRENT_DIR);
+    if (!d) {
+        return 0;
+    }
+    int num_files = 0;
+    fs_dir_entry *entry;
+    struct stat file_info;
+    while ((entry = fs_dir_read(d)) && num_files < max_files) {
+        const char *name = dir_entry_name(entry);
+        if (stat(name, &file_info) != -1) {
+            int m = file_info.st_mode;
+            if (S_ISDIR(m) || S_ISCHR(m) || S_ISBLK(m) || S_ISFIFO(m) || S_ISSOCK(m)) {
+                continue;
+            }
+        }
+        if (file_has_extension(name, extension)) {
+            strncpy(file_list[num_files], name, FILE_NAME_MAX);
+            file_list[num_files][FILE_NAME_MAX - 1] = 0;
+            ++num_files;
+        }
+        dir_entry_close_name(name);
+    }
+    fs_dir_close(d);
+    *file_count = num_files;
+    return 1;
+#endif
+}
+
+int platform_file_manager_set_base_path(const char *path)
+{
+    if (!path) {
+        log_error("set_base_path: path was not set. Julius will probably crash.", 0, 0);
+        return 0;
+    }
+#ifdef __ANDROID__
+    return android_set_base_path(path);
+#else
+    return chdir(path) == 0;
+#endif
+}
+
+static int correct_case(const char *dir, char *filename)
+{
+    // Note: we do not use the _w* variants for Windows here, because the
+    // Windows filesystem is case insensitive and doesn't need corrections
+    DIR *d = opendir(dir);
+    if (!d) {
+        return 0;
+    }
+    struct dirent *entry;
+    while ((entry = readdir(d))) {
+        if (string_compare_case_insensitive(entry->d_name, filename) == 0) {
+            strcpy(filename, entry->d_name);
+            closedir(d);
+            return 1;
+        }
+    }
+    closedir(d);
+    return 0;
+}
+
+static void move_left(char *str)
+{
+    while (*str) {
+        str[0] = str[1];
+        str++;
+    }
+    *str = 0;
+}
+
+const char *platform_file_manager_case_correct_file(const char *filepath)
+{
+    // Android already checks for proper case in the Java side, so if it was not found before, it won't be now
+#ifndef __ANDROID__
+    static char corrected_filename[2 * FILE_NAME_MAX];
+
+    strncpy(corrected_filename, filepath, 2 * FILE_NAME_MAX);
+    corrected_filename[2 * FILE_NAME_MAX - 1] = 0;
+
+    char *slash = strchr(corrected_filename, '/');
+    if (!slash) {
+        slash = strchr(corrected_filename, '\\');
+    }
+    if (slash) {
+        *slash = 0;
+        if (correct_case(".", corrected_filename)) {
+            char *path = slash + 1;
+            if (*path == '\\') {
+                // double backslash: move everything to the left
+                move_left(path);
+            }
+            if (correct_case(corrected_filename, path)) {
+                *slash = '/';
+                return corrected_filename;
+            }
+        }
+    } else {
+        if (correct_case(".", corrected_filename)) {
+            return corrected_filename;
+        }
+    }
+#endif
+    return 0;
+}
+
+int platform_file_manager_file_exists(const char *filename)
+{
+#ifdef __ANDROID__
+    return android_check_file_exists(filename);
+#else
+    return NULL != platform_file_manager_case_correct_file(filename);
+#endif
+}
+
+int platform_file_manager_remove_file(const char *filename)
+{
+#ifdef __ANDROID__
+    return android_remove_file(filename);
+#else
+    return remove(filename) == 0;
+#endif
+}
+
+#if defined(__vita__)
+
+FILE *platform_file_manager_open_file(const char *filename, const char *mode)
+{
+    char *resolved_path = vita_prepend_path(filename);
+    FILE *fp = fopen(resolved_path, mode);
+    free(resolved_path);
+    return fp;
+}
+
+#elif defined(_WIN32)
+
+FILE *platform_file_manager_open_file(const char *filename, const char *mode)
+{
+    wchar_t *wfile = utf8_to_wchar(filename);
+    wchar_t *wmode = utf8_to_wchar(mode);
+
+    FILE *fp = _wfopen(wfile, wmode);
+
+    free(wfile);
+    free(wmode);
+
+    return fp;
+}
+
+#elif defined(__ANDROID__)
+
+FILE *platform_file_manager_open_file(const char *filename, const char *mode)
+{
+    int fd = android_get_file_descriptor(filename, mode);
+    if (!fd) {
+        return NULL;
+    }
+    return fdopen(fd, mode);
+}
+
+#else
+
+FILE *platform_file_manager_open_file(const char *filename, const char *mode)
+{
+    return fopen(filename, mode);
+}
+
+#endif
