+++ test/CMakeLists.txt
@@ -38,6 +38,7 @@ add_executable(autopilot
     stub/sound_device.c
     stub/ui.c
     stub/video.c
+    ${PROJECT_SOURCE_DIR}/src/platform/file_manager.c
     ${TEST_CORE_FILES}
     ${TEST_BUILDING_FILES}
     ${CITY_FILES}
